+++ src/components/Listedcard/Listedcard.jsx
@@ -18,42 +18,42 @@ const Listedcard = (props) => {
   const links = Array.isArray(getlink) ? getlink.slice(0, 1) : [];
 
   const handleLink = (data) => {
-    nav(`/${data}`, { state: { title: data } });
+    nav(`/card-page/${data}`, { state: { title: data } });
   };
 
   return (
     <section id={section} className="">
-      <div className="flex   md:mt-1">
+      <div className="flex flex-col  md:mt-1">
         <h1 className="text-4xl  mx-2 font-bold">{t(title)}</h1>
       </div>
-      <div className="w-full gap-6 mr-12">
+      <div className="w-full gap-6 mr-12 gap-3">
         {listItems
           ? listItems.map((item, index) => (
               <div
                 key={index}
-                className="w-full h-auto mb-6 border border-[#013A63]/15 bg-[#013A63]/5 rounded-md p-4"
+                className="w-full h-auto  rounded-md gap-3 py-2 cursor-pointer"
+                onClick={() => handleLink(item.title)}
               >
-                <article className="group overflow-hidden rounded-lg border border-gray-100 bg-white shadow-sm ">
-                  <img
-                    alt=""
-                    src={item.images || item.image}
-                    className="h-56 w-full rounded-lg object-cover  "
-                  />
-                  <div className="p-8 bg-[#013A63]/5">
+                <article className="flex group overflow-hidden rounded-lg border border-gray-100 p-2 bg-[#013A63]/5 shadow-sm ">
+                  <div className="flex">
+                    <Images
+                      alt=""
+                      src={item.images || item.image}
+                      className="h-40 w-52 w-full rounded-lg object-cover  "
+                    />
+                  </div>
+                  <div className="p-8  w-full ">
                     <div className="flex items-center justify-between ">
                       <h3 className="text-lg font-medium text-[#013A63] fira-sans-bold">
                         {item.title}
                       </h3>
-                      <span className="whitespace-nowrap rounded-full bg-transparent border border-[#013A63] px-3.5 py-1.5 fira-sans-condensed-bold text-sm text-[#013A63]">
-                        <Dateformat dateString={item.created_at} />
-                      </span>
                     </div>
                     <p className="mt-2 line-clamp-3 text-sm/relaxed text-gray-500 fira-sans-condensed-regular">
                       {item.description}
                     </p>
                     <a
                       className="group mt-4 inline-flex items-center gap-1 text-sm text-[#013A63] fira-sans-condensed-bold"
-                      href={item.title}
+                    onClick={() => handleLink(item.title)}
                     >
                       Find out more
                       <span
@@ -69,14 +69,6 @@ const Listedcard = (props) => {
             ))
           : ""}
       </div>
-      {/* <div className="mt-5 text-center">
-          <a
-            href={`/${links}`}
-            className=" p-2  text-gray-700 border  hover:bg-blue-500 hover:text-gray-100 rounded-lg"
-          >
-            View More {`>>`}
-          </a>
-        </div> */}
     </section>
   );
 };
