+++ src/pages/Homeview.jsx
@@ -1,19 +1,18 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
-import Businessview from "./Homeview/Businessview/Businessview";
 import { HeroBanner } from "./Homeview/HeroBanner/HeroBanner";
 
 function Homeview({ data }) {
   const { t } = useTranslation();
 
   return (
-    <div className='w-full'>
-      <HeroBanner blogData={data.blogData} />
-      {/* <Herobanners /> */}
-      <Businessview
+    <div className="w-full">
+      <HeroBanner
+        blogData={data.blogData}
         navbar={data.navbar}
         businessCarousel={data.businessSettings}
       />
+   
     </div>
   );
 }
